[ADD] : 페이지 리스트 인덱스를 서비스 별로 지정할 수 있는 기능 추가
